Add assert feature at important points
There are several classes that require the assert feature to ensure
that there are no runtime errors. In some cases, the asserts seem trivial
but are included for comprehensiveness.

Moving forward, the project will liberally include the use of assert features
for the safety features they provide. Whilst costing little overhead and performance
costs.